Déplacement & Rotation de la créature

+++ testAllegro/src/map/Position.cpp
@@ -4,7 +4,7 @@ Position::Position() : Position(0, 0){
 
 }
 
-Position::Position(int x, int y){
+Position::Position(double x, double y){
 	this->x = x;
 	this->y = y;
 }
@@ -14,23 +14,23 @@ Position::Position(Position& copy){
 	y = copy.y;
 }
 
-int Position::getX(){
+double Position::getX(){
 	return x;
 }
 
-int Position::getY(){
+double Position::getY(){
 	return y;
 }
 
-void Position::setX(int x){
+void Position::setX(double x){
 	this->x = x;
 }
 
-void Position::setY(int y){
+void Position::setY(double y){
 	this->y = y;
 }
 
-void Position::setPosition(int x, int y){
+void Position::setPosition(double x, double y){
 	this->x = x;
 	this->y = y;
 }
